fix(terminal): partially revert previous mappings
Commit 4efd0e2bea2a387d6fcdea8cf36a62049e8bafed introduced some changes
to terminal mappings. I tested in a new LazyVim installation and
`sidekick.nvim` is not affected by `<C-/>` any more.

The mappings `hide_underscore` and `hide_slash` don't let user to create
new terminals with `2<C-/>`, `3<C-/>` and so on. Therefore, I believe
it would be a better default to revert these 2 mappings explicitly.